- Refactor types - Move data fetching & formatting functions to GalleryCategoryFilter.tsx level
